refactor(questions): Move pass rate logic to Question
