+++ src/Entity/Question.php
@@ -177,6 +177,44 @@ class Question
     }
 
     /**
+     * Get the pass rate of the question.
+     *
+     * @return float the pass rate of the question
+     */
+    public function getPassRate(): float
+    {
+        $totalAttemptCount = $this->getTotalAttemptCount();
+        if (0 === $totalAttemptCount) {
+            return 0;
+        }
+
+        return round($this->getTotalSolvedCount() / $totalAttemptCount * 100, 2);
+    }
+
+    /**
+     * Get the total number of attempts made on the question.
+     *
+     * @return int the total number of attempts made on the question
+     */
+    public function getTotalAttemptCount(): int
+    {
+        return $this->getSolutionEvents()->count();
+    }
+
+    /**
+     * Get the total number of times the question has been solved.
+     *
+     * @return int the total number of times the question has been solved
+     */
+    public function getTotalSolvedCount(): int
+    {
+        return $this->getSolutionEvents()
+            ->filter(
+                fn (SolutionEvent $solutionEvent) => SolutionEventStatus::Passed === $solutionEvent->getStatus()
+            )->count();
+    }
+
+    /**
      * @return Collection<int, SolutionEvent>
      */
     public function getSolutionEvents(): Collection
